\name{GenerateRandomLetter-package}
\alias{GenerateRandomLetter-package}
\alias{GenerateRandomLetter}
\docType{package}
\title{
\packageTitle{GenerateRandomLetter}
}
\description{
\packageDescription{GenerateRandomLetter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GenerateRandomLetter}
\packageIndices{GenerateRandomLetter}
~~ This package contains the function ralpha, which returns ~~
~~ a random letter of the English alphabet ~~
}
\author{
\packageAuthor{Paul Davies}

Maintainer: \packageMaintainer{Paul Davies}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
